feat: se agrega carpetas faltantes del proyecto

+++ src/main/java/com/cristianmanuel/Kinalapp/repository/ClienteRepository.java
@@ -3,5 +3,7 @@ package com.cristianmanuel.Kinalapp.repository;
 import com.cristianmanuel.Kinalapp.entity.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+// Repositorio JPA para la entidad Cliente. Proporciona métodos CRUD básicos.
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+    // Long es el tipo de la clave primaria (dpi_cliente)
 }
