+++ src/main/java/com/cristianmanuel/Kinalapp/repository/DetalleVentaRepository.java
@@ -3,5 +3,7 @@ package com.cristianmanuel.Kinalapp.repository;
 import com.cristianmanuel.Kinalapp.entity.DetalleVenta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetalleVentaRepository extends JpaRepository<DetalleVenta, Integer> {
+// Repositorio para DetalleVenta. Hereda métodos como save(), findAll(), deleteById().
+public interface DetalleVentaRepository extends JpaRepository<DetalleVenta, Long> {
+    // Long corresponde al tipo de codigoDetalleVenta
 }
