+++ src/main/java/com/cristianmanuel/Kinalapp/repository/ProductoRepository.java
@@ -3,5 +3,7 @@ package com.cristianmanuel.Kinalapp.repository;
 import com.cristianmanuel.Kinalapp.entity.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+// Repositorio para la entidad Producto.
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
+    // Long = codigoProducto
 }
