+++ src/main/java/com/cristianmanuel/Kinalapp/repository/VentasRepository.java
@@ -3,5 +3,7 @@ package com.cristianmanuel.Kinalapp.repository;
 import com.cristianmanuel.Kinalapp.entity.Ventas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VentasRepository extends JpaRepository<Ventas, Integer> {
+// Repositorio para la entidad Ventas.
+public interface VentasRepository extends JpaRepository<Ventas, Long> {
+    // Long = codigoVenta
 }
